refine: using new Prisma generator
Declared in the schema, the generator is "prisma-client" instead of "prisma-client-js" and is in preview stage. This is going to be the standard generator soon. Doing this changes the generated client output in many small ways, but using Prisma and the syntax of queries and mutations will remain the same.

I had to update some type imports in my Update Profile server action due to the new generated client folder structure